testing/fastapi oracledb react/backend/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
oracledb
